// Capstone Project/src/app/pages/landlord/room-create-page/screen/types.ts
export enum hostRoomCreateEnum {
  CREATE_ROOM_REQUEST = 'APP/HOST_ROOM_CREATE/CREATE_ROOM_REQUEST',
  CREATE_ROOM_SUCCESS = 'APP/HOST_ROOM_CREATE/CREATE_ROOM_SUCCESS',
  CREATE_ROOM_FAILURE = 'APP/HOST_ROOM_CREATE/CREATE_ROOM_FAILURE',


  CHECK_ROOM_UNIQUE_REQUEST = 'APP/HOST_ROOM_CREATE/CHECK_ROOM_UNIQUE_REQUEST',
  CHECK_ROOM_UNIQUE_SUCCESS = 'APP/HOST_ROOM_CREATE/CHECK_ROOM_UNIQUE_SUCCESS',
  CHECK_ROOM_UNIQUE_FAILURE = 'APP/HOST_ROOM_CREATE/CHECK_ROOM_UNIQUE_FAILURE',

  LOADING_REQUEST = 'APP/HOST_ROOM_CREATE/LOADING_REQUEST',
  LOADING_SUCCESS = 'APP/HOST_ROOM_CREATE/LOADING_SUCCESS',
  LOADING_FAILURE = 'APP/HOST_ROOM_CREATE/LOADING_FAILURE',

  CLEAR_HOST_ROOM_CREATE_STATE = 'APP/HOST/HOST_ROOM_CREATE_STATE/CLEAR_HOST_ROOM_CREATE_STATE',
}